Bugs in the Brian module, due to the distintion between a compressed connection matrix and not. Both do not have the same arguments. Now, the default is to compress a connection after it has been created. This solution works, and allow to access the connections before or after a run command in the same manner.
git-svn-id: https://neuralensemble.org/svn/PyNN/trunk@840 1a7b9c28-352d-0410-8541-c879d6df8dcf

+++ src/nest/__init__.py
@@ -404,7 +404,7 @@ class Projection(common.Projection):
         logger.debug("--- Projection[%s].__saveConnections__() ---" % self.label)
                 
         if gather == False or rank() == 0:
-            lines       = numpy.array(lines)
+            lines       = numpy.array(lines, dtype=float)
             lines[:,2] *= 0.001
             if compatible_output:
                 lines[:,0] = self.pre.id_to_index(lines[:,0])
